build: configure lint-staged and pre-commit hook
Adds a Husky pre-commit hook that runs lint-staged, which has been configured to run 'lint' and
'format' on the file types specified in .lintstagedrc. This helps the developer by only running the
lint and format scripts on files they have changed and staged, and keeps unlinted and unformatted
commits out of the code base.

Resolves: #26