fix: correct problems with Typescript definitions and JSDoc comments
This might fix the problem in #43. I added @ts-check back in, changed the index.d.ts to 'declare
class' instead of 'export class', and I am not getting any errors in VS Code anymore with the built
in Typescript checker. However, I am not actually running Typescript, so I do not know if this will
solve the problem once and for all or not. Additionally, I indentified some issues in the API
documentation and have improved those.